Set things up to run a server in the background
Extend the server code so that we can run a server process in the
background. Since we can't change the logging after it is setup, we exec
a new copy of the server in foreground mode once we have a daemon
running, and use an environment variable to setup the logging to go to a
file in the state directory.

We now also have started using the XDG paths for storing things.