Rebase seems to be working now
It looks like the issue was that you *do* need to commit the changes for
a rebase, but when a commit is empty then you don't need to commit that
change (since there is no change). So far I can't see any way to know
that you don't need to commit anything other than trying and then
catching the "Applied" error code that you get back.