Start working on the shell integration side
Start working on the mechanisms to integrate the tool into the shell.

We are going to take the fairly standard route of using `eval "$(vellum
init <shell>)"` as the method. So we are going to embed the necessary
setup scripts into the binary.

Starting with zsh because the integration is a lot easier there, since
the hooks that we need to connect things up are already there as
standard.