Start working on more complete history storage
Currently the way that we store history is very naive. We are basically
reading and writing complete files every time. And just storing the data
using a simple JSON encoding.

We want to change this so that we are encypting the contents before
writing to disk, or to git (since command history could have all kinds
of secrets in it that shouldn't be stored in plaintext).

Since encrypting the data is going to make it binary, if we kept the
approach of re-writing the files every time, then we would be basically
creating complete file diffs for git - which is rather inefficient for
the history storage.

So in addition to wanting to add encryption, we want to add a different
way of writing the data, where we only add new data when writing, so
that the git diffs are always just additions of new data, without
changes to existing data.

The basic plan is to write encrypted chunks, which contain one or more
commands, with timestamps so that we can track which things are new
since the last read/write operation.