Start working on adding the encryption layer
We want to store the history chunks encrypted, so start working on
setting that up. We want to serialise the Chunk data, and encrypt it,
then store that encrypted data with unencrypted metadata (so that we can
read the metadata without needing to decrypt the data).

Create a new EncryptedChunk type to do this, with the idea that we will
be serialising/deserialising Vec<EncryptedChunk> to disk.

We also start using MessagePack instead of JSON, since JSON is a
terrible way to store binary data, and most of the data we are writing
is now going to be encrypted, it makes more sense to use a binary
serialisation format.