Tweak things to add a combined sync method
Tweak things around a bit so that we can have a combined sync method.
This makes the git interaction a little bit more efficient.

Also, we should start with the last_read sufficiently far in the past
that we pick up all existing history data on the first read.

There was also nothing making sure that the directory to represent a
host actually existed, so add that too.