Add a command to output an encryption key
Add a sub-command to init that outputs a generated encryption key. This
allows the use of `vellum init key` to generate an encryption key that
can be used for $VELLUM_KEY.

Also add a check to the server startup, so that we refuse to start a
server if the $VELLUM_KEY value isn't set to a usable value.