Move syncer into History
Starting to look at moving over to using History, and I realised that
the only thing State had that History didn't was it owned the Syncer. So
update History to own the Syncer instead having it passed in, so that
History can just fully replace State.

This did run into an issue with the GitUpdate type holding a reference,
however it turned out that was an overly complicated way to do things
anyway. Much simpler to move the finish method into Git, and pair the
new method with refresh, instead of holding onto some state that didn't
actually hold onto anything interesting anyway.