Fix some little bugs in the history code
When reading, we would never read our own history, even on startup. So
now if we have no history at all we will read our own history on the
assumption that the server has just started.

When writing, the filter was backwards so we were only writing old
chunks, instead of only writing new chunks.

Things look to be working as expected now. The git history is useless,
because we are writing binary files, but the storage should be efficient
since we are only ever appending data to files, so git's delta encoding
should only be needing to store the new data.