Implement background sync thred for server
The server will now run a background thread to perform a sync on a
ticker, based on an interval specified in the config.

This has revealed an issue though. We aren't tracking reads properly. We
can't assume that the files are up to date when we read them, as the
other server may not have pushed it's changes yet. So we will be reading
things with a delay, and with different delays for each server.

So instead of a single last_read timestamp that is set to the current
time, we need to look at the last entry timestamp we have seen for each
host.