Change the way we track read status
Instead of maintaining a single last_read timestamp, for each host we
use the timestamp of the latest chunk as the last_read timestamp. This
ensures that we don't care about the delay introduced by the difference
between when a chunk gets created, and when we get to see it.

We still ignore anything we see that we have already read, but we no
longer incorrectly ignore things that were written locally before our
last read, but only synced after.