Store all entries with a session ID
The client gets a session ID from an environment variable, and it is
included with the command to be stored.

The idea here is that we can then ask for the history to be filtered to
only things from the current session. This should make using the up
arrow less confusing, as we can restrict the up arrow to commands from
the current session, whilst allowing search access to the full history.