Truncate the pid file after opening
Since we don't open the pid file with truncate, we need to manually
truncate it after opening. Otherwise if we write a smaller value than
was previously in the file the end of the previous value will be left
and reads will return the wrong value.

We don't truncate as part of opening because that would happen before we
find out that another process has the file locked - so we would have
truncated the pid of another server whilst it was still running.