Move syncer out of History
By having the Syncer owned by History we had to lock the History for the
entire duration of the sync. Since the bulk of the time is spent
interacting with git, and that takes hundreds of milliseconds to
seconds, this can cause client calls to stall for a noticable amount of
time. So we move the Syncer out in the Server, so that we can lock the
Syncer separately, and only lock the History for the time taken to
actually do the History sync work, which is much faster.

This should hopefully resolve the issue with the clients getting blocked
whilst the server talks to git.