Add ability to mark entries as deleted
You can now mark entries as deleted using the `vellum delete` command.
The api actually provides the ability to update an entry, where setting
the entry to an empty command will delete it.

The updates don't actually change anything in the chunks. Chunks remain
append only, as does the on-disk format. However, now when we rebuild
the merged Vec we collect entries for the same ID and collapse them,
then keep the metadata from the original entry, but the cmd from the
most recent. After that we filter out any entries with an empty cmd
(thus deletion is just storing an update with an empty cmd).

Whilst this only allows removing entries from the history, and doesn't
remove entries from disk or git, the plan is to follow this up with a
`vellum rebuild` command that will rebuild the on-disk files from the
merged data, thus forgetting any original entries and only keeping the
latest state. Once the files have been rebuilt then this command should
create a new git history that only contains the new files.