Add a command to edit history
This new command retreives a chunk of history based on the filter
options provided, and then presents it in an editor. The lines can then
be edited or removed to modify or delete those entries in the history.

As with the delete command this command records the changes, without
acutally changing files stored on disk / in git. To fully change the
stored history then `vellum rebuild` must be run after `vellum edit`.