Make edit behaviour more controllable
Now edit will ask for confirmation by default, and can be run in a
dry-run mode. This gives the user more opportunity to ensure that the
changes are what they want before comitting to them.

We also have options so that the user can say "just apply my changes
without telling or asking me" if they want.