Implement repo locking using a git tag
We can now lock and unlock the repo by creating or deleting a tag in the
git repo. If the refresh sees that the repo has been locked, then it
will run in a loop until the repo is unlocked, or it times out waiting.

One issue currently is that the background sync will queue up all the
missed syncs rather than catching up with a single sync ...