Build out the syncer API for when locked
We want to be able to work with the locked repo to do the rebuild, so
add some more methods to the return value from lock that parallel the
unlocked versions. Also rename the return type trait to better reflect
its purpose, it is to provide an API that should only be used when the
repo is locked, not to act like a MutexGuard.

Currently the push function doesn't actually work as it should, it's
just the same as a regular push with a different commit message, but it
should also flatten the git history.