Implement a new commit to flatten history
As part of the rebuild we now create a new parent-less commit that
contains the updated files, and update the local branch to point at that
new commit.

We then try to push the new branch state, and it fails because it is not
a fast-forward. Which is expected, the next step is to make the push a
force-with-lease style push.