Actually force push the changes
Now that we have the "with-lease" part in place, we can update the
refspec so that we are actually force pushing the new branch.

Only problem now is that other servers are not too happy with the new
branch (though manually running `git pull -r` updates the repo just
fine.