Change the way that we run rebase
Now we keep track of the original oid of the upstream branch when we see
it has changed. This means that when we run the rebase we can use the
old upstream oid as the starting point, and the current upstream branch
as the destination.

This should makes pretty much all of the rebases that we do trivial,
since everything we do apart from a rebuild is additive, and a rebuild
locks the repo to prevent having to merge local changes with a rebuild.