Make push_changes a little more robust
Previously we were detecting if a push was needed based on whether
commit had made a new commit, but this didn't work if a previous push
had failed for some reason (e.g. conflict over updating refs). Change
the code so that we instead look at if our branch is ahead of the
uptream branch to see if we need to push.

Also tidy up a bunch of code to use some of the new functions that were
added as part of figuring out if we had any local changes.