Add ensure option to server command
Add a flag to the server command that tells it that we just want to be
sure the server is running.

This allows running `vellum server --ensure` in a shell setup file or
similar to make sure that the server is running, without worrying about
errors when the server has already been started.