Tweak the way that ping works
Instead of ping having a fixed duration, and a wait flag, it now takes
an Option<Duration> so that we can wait for different durations in
different uses. In addition the established connection is returned on
success (and ensure_ready is similarly tweaked) - so that the client
command can simply use that connection instead of disconnecting and then
creating a new connection.