Tolerate network failures on fetch
If we receive a network error when calling fetch, then just log a
warning, and ignore it. This is because we still want the sync to
continue to update the local git repo. That way if we are offline we'll
just build up local commits until we can connect to the remote repo
again, instead of failing and holding the changes in ram only.