Make some tweaks to the history command
Change the default number of lines of history printed, so that only the
most recent 10 commands are shown by default. Also tweak the processing,
so that if a negative index is too large we simly take it as meaning
everything in the list.

Also add an option to show the index numbers on the output.