Rework how indexing works for history command
Previously the indexes shown, and used to select entries to be shown
were generated after the filtering was applied. However, this means that
the indexes are unstable, and do not always refer to the same command.

So now the indexes are generated before the filtering is applied. For
the negative values though, they are calculated *after* filtering. This
makes the selecting a bit harder to apply, but it behaves more logically
for the user, as "-10" means the most recent 10 entries, regardless of
what the filtering selects.