Add save and load commands
Add commands to save and load commands to/from a JSON file. The primary
purpose of these new commands is to deal with issues with sync. If the
sync breaks then the save command can extract the history that hasn't be
stored to disk, and the load command can load them into a new instance
of the server that hopefully can sync those commands.