Keep a copy of the active chunk on disk
This is the first step in trying to make the server less significant.

Currently until the server does a sync, the history is stored only in
memory. This can be a problem if the sync breaks, as the history can
build up in memory, and you have to manually extract it before stopping
the server, and then manually load it into the new instance.

This update is the beginning of changing that. Now the active chunk is
also held on disk in addition to in memory. This means that if the
server is stopped before it runs a sync that all un-synced history
should still be on disk. Once the code is updated to read in the active
chunk on startup then a stuck sync can be fixed by simplying restarting
the server, as it will reload the unsynced active chunk data, and
hopefully sync it using the now-unstuck background sync.

The other change to improve robustness that should follow is to have the
server automatically restart if the background sync hasn't run for a
while (not run and failed, but not run). Since libgit2 doesn't have a
timeout mechanism currently, there isn't a simple way to ensure that git
sync operations don't get stuck, but having a time based restart that
happens if the background sync hasn't run in a long time should
workaround the issue, when combined with the persistent active chunk
that we are starting to add here.