Fix sync issues from writing active chunk
The code that was added to write out the active chunk introduced some
issues with the sync code because it introduced the idea of an empty
active chunk. Previously an active chunk would only be created if
something was being added to it, but now the code creates an empty
active chunk if there isn't one so that it can write it to disk.

This was confusing the sync code, as it was previously assuming that all
chunks had entries. So it was continously syncing the empty chunks to
git. It was also causing new empty active chunks to be created all the
time, as "last_write" was being updated every time, even if nothing had
actually been written.

These changes address those issues - though a better fix may be to tweak
things so that we stop producing the empty active chunk just to write to
disk.