write_active no longer causes empty active chunks
We now write an empty file if there is no active chunk, instead of
creating an empty active chunk just to write it to disk.

This means that the workaround of having write return if anything was
written is not needed any more, and last_write can always be updated
again.