Add a watchdog background thread
Now if we start the sync background thread, we also start an extra
watchdog background thread. This extra thread just monitors how long it
has been since the last sync. We don't pay attention to the result of
the sync, just that it has completed.

If the time since the last sync is more than the configured timeout,
then the watchdog will exit the server. This shouldn't lose any data, as
that should all be on disk, and will be reloaded on restart. However,
exiting the process should hopefully unstick any stuck sync process.