Run initial sync in background
Previously the server would run a sync as part of startup before the
actual server was running. This could cause an large delay before the
server would respond to pings. Change the startup sequence so that we
instead run a initial sync in the background.

This means that the history contents may be slightly out of date until
that initial sync completes, but that is preferable to the client
failing to store a command because the server doesn't respond.