Start trying to add path to Entry
Adding the path to Entry is made more complicated by the fact that
rmp_serde is using a list to encode structs. This means that we can't
add an extra field an expect to decode old data. So this starts the work
of introducing a version field, which is stored in the high byte of the
length.

Currently we are still working with the same data structures, and don't
do anything with the version. The next step is to use the version to
change the chunk decoding and support v1 and v0 chunks, with conversion
of v0 chunks into v1 chunks with empty path strings in the entries.