Introduce the concept of an API verison
Add a payload to the Pong message that carries an API version. This will
let the client detect when the server is running an incompatible API
version and return an appropriate error.

Since changing the API version prevents the client from talking to the
server at all it should only be changed if the API is changed in a way
that could cause miscommunication instead of simple errors.