Start passing the current directory to the server
The store command now looks up the current directory, and passes that
path through to the history code in the server. This path is then stored
in the updated Entry and is available to the user in JSON output.

There are still some things to sort out:

 * What do we do about editing history, how do we show/edit paths
 * What do we do about updates, do we store a new path?
 * What else needs updating to handle paths?
 * How do we stop old clients blowing up with the changed file format

The last point is quite important. Since we are potentially running on
multiple machines we have to assume that the new format files we are
writing might be read by and old version of the server. Since the old
code wasn't expecting a version in the top-byte of the length then it
will read it as a very large length, and try to read far too much data.
The length will almost certainly be larger than the file, so it will
probably just error out with an unexpected EOF. At the very least this
behaviour needs to be checked. Ideally updating one host in a
distributed setup wouldn't break all the other hosts, but just degrade
gracefully (e.g. ignoring the new data from the new server).