Start splitting out the history store code
Previously the code to manage the storage of history, and the in-memory
history data was all in the same module. This was a bit unweidly, and
adding support for versioning was making it worse. So start splitting
the storage code out into its own sub-module. Currently only some of the
storage code has been moved, and the imports in the store module haven't
been tidied up (as some more are likely to be used once all the code is
moved).