Start moving chunk reading into store
This moves the actual file reading part of read_host into store.

It also tidies up the use statements in store, since the remaining
missing imports aren't related to reading or writing the history files.

EncryptedChunk is also fully private again, now that all of the code
that uses it is in store.