Move the last bit of reading code into store
Now the code that figures out which paths need reading for which hosts
lives in store. The connection between this and actually reading the
hosts stays in history, because it wants to do some filtering over which
hosts get loaded, and does some extra work for each host that gets
loaded.