Ignore chunks of unknown version
Instead of erroring when we see a chunk of an unknown version, ignore it
and log a warning instead.

This provides better support for rolling out updates, as other servers
that are sharing the same repo but which haven't been updated will keep
working, they will just start ignoring updates from the updated servers.
Instead of erroring and basically refusing to load any more history.

If this code had alredy been in place, then adding support for storing
paths would have been a lot simpler.