Fix writing of state file
If we aren't appending to the file, then we need to explicitly set
write. If we are appending then writing is implied, so previously we
didn't need to explicitly set write.

So when writing the state file was moved into HistoryFile, the writing
actually stopped working, because we didn't open the file for writing.