Another fix for writing the state file
When we are writing the state file, we are relying on truncation of the
file to remove old data. So change the create function so that it sets
either append (for writing normal chunk files) or truncate (for writing
the state file). Without this old data wasn't being cleared from the
state file, causing a read to read random data as a chunk length and
crash the application.