Write chunk files into new location
Now all the pieces are in place to write the chunk files to a new
loation, whilst still reading from the original location.

This will allow updating vellum without other servers running into
issues with the new file format. Since the old format files can be read
by the new code we currently don't have a mechanism to tell read_host
the difference between old and new, as we don't need it.